% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly_numbers.R
\name{divisors}
\alias{divisors}
\title{Calculate the divisors of a number}
\usage{
divisors(num)
}
\arguments{
\item{num}{the integer number}
}
\value{
A Vector of Numbers
}
\description{
Returns a vector with all the divisors
}
\examples{
\dontrun{
divisors(12)
divisors(1400)
}
}
