% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_loop.R
\name{factorial_loop}
\alias{factorial_loop}
\title{Calculates the factorial of a number}
\usage{
factorial_loop(n)
}
\arguments{
\item{n}{the integer to calculate the factorial}
}
\value{
the factorial
}
\description{
Calculates the factorial using a loop.
}
\examples{
\dontrun{
factorial_loop(10)
x <- factorial_loop(7)}

}
