% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly_numbers.R
\name{friendly_numbers}
\alias{friendly_numbers}
\title{Calculates whether two integer numbers are friendly numbers}
\usage{
friendly_numbers(num1, num2)
}
\arguments{
\item{num1}{The first integer number}

\item{num2}{The second integer number}
}
\value{
A Logical TRUE/FALSE
}
\description{
Returns a TRUE/FALSE to determine whether two integer numbers are friendly
}
\examples{
\dontrun{
friendly_numbers(12, 16)
friendly_numbers(220,284)
}
}
