% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly_numbers.R
\name{perfect_number}
\alias{perfect_number}
\title{Determines whether a number is perfect or not}
\usage{
perfect_number(num)
}
\arguments{
\item{num}{the integer number}
}
\value{
TRUE or FALSE
}
\description{
Returns a logical TRUE/FALSE
}
\examples{
\dontrun{
perfect_number(12)
perfect_number(6)
}
}
